// src/shared/api/model/wordFilterTypes.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 1.0.0
 */

/**
 * An enumeration.
 */
export type WordFilterTypes = typeof WordFilterTypes[keyof typeof WordFilterTypes];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WordFilterTypes = {
  checked: 'checked',
  unchecked: 'unchecked',
  all: 'all',
} as const;